+++ helper-scripts/deleteStacksZx/functions/mocks/mockFunctions.ts
