+++ helper-scripts/deleteStacksZx/helpers/formatting.ts
