+++ helper-scripts/deleteStacksZx/protectedStacks.ts
